Updated magnalister build version 8706 to latest modified-shop compatible build version 8800
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@16286 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ includes/external/magnalister/php/modules/check24/prepare.php
@@ -57,7 +57,15 @@ class Check24Prepare extends MagnaCompatibleBase {
 		} else if (!empty($aProductIDs)) {
 			$oProductSaver->saveMultipleProductProperties($aProductIDs, $itemDetails);
 		}
-		
+
+		if (count($oProductSaver->aErrors) > 0) {
+			$_POST['prepare'] = 'prepare';
+			foreach ($oProductSaver->aErrors as $sError) {
+				echo '<div class="errorBox">'.$sError.'</div>';
+			}
+			return;
+		}
+
 		MagnaDB::gi()->delete(TABLE_MAGNA_SELECTION, array(
 			'mpID' => $this->mpID,
 			'selectionname' => $this->prepareSettings['selectionName'],
